% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_dfs.R
\name{compare_dfs}
\alias{compare_dfs}
\title{compare_dfs}
\usage{
compare_dfs(df1, df2, metric)
}
\arguments{
\item{df1}{First data.frame for comparison}

\item{df2}{Second data.frame for comparison}

\item{metric}{Function name or anonymous function that returns a value that can be compared.}
}
\value{
A named logical vector indicating if the metric evaluated to the same result for each column.
}
\description{
Compares two data.frames using a function that when applied to each row, returns a value that can be compared.
}
