% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concat_excel_sheets.R
\name{concat_excel_sheets}
\alias{concat_excel_sheets}
\title{concat_excel_sheets}
\usage{
concat_excel_sheets(path, sheets = "all", quiet = FALSE, ...)
}
\arguments{
\item{path}{Path to the .xls or .xlsx document.}

\item{sheets}{A numeric vector of indices or a character vector of names for the sheets to use.}

\item{quiet}{Silences printing progress to the console.}

\item{...}{(Optional) To pass additional arguments to read_excel().}
}
\value{
A tibble containing the merged observations from the selected sheets. All column data types are left as 'character' to facilitate when binding multiple Excel files.
}
\description{
Concatenates selected sheets within an Excel file. It does not infer data-types to facilitate binding sheets Use readr::type_convert() to guess column types. Use ... to pass additional arguments to read_excel().
}
