% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/establish_db_connection.R
\name{establish_db_connection}
\alias{establish_db_connection}
\title{Title}
\usage{
establish_db_connection(database = "NRADWH")
}
\arguments{
\item{database}{The name of the database connection to establish. It must match the name given to 'setup_database_access()'}
}
\value{
An object returned by DBI::dbConnect() which can be used for making subsequent queries against the defined database.
}
\description{
Title
}
