% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_asycuda_import_export.R
\name{fetch_asycuda_import_export}
\alias{fetch_asycuda_import_export}
\title{Get ASYCUDA Import-Export Data}
\usage{
fetch_asycuda_import_export(
  db_con,
  cols = NULL,
  reg_date = NULL,
  dec_type = NULL
)
}
\arguments{
\item{db_con}{An active DBIConnection object, often created by 'establish_db_connection()'}

\item{cols}{(Optional) A vector or list of columns to include in the subset.}

\item{reg_date}{(Optional) A vector or list where the first element is the start-date and second element the end-date.}

\item{dec_type}{(Optional) A vector or list of Tax Types to filter on. Options are:}
}
\value{
A data.frame containing all the matching observations in the Data Warehouse.
}
\description{
Get ASYCUDA Import-Export Data
}
\examples{
# Establish database connection
NRADWH <- establish_db_connection(database = "NRADWH")

# Fetch Import-Export results for 2022
fetch_asycuda_import_export(db_con = NRADWH, reg_date = c("2022-01-01", "2022-12-31"))

# Fetch results for IM4 and IM7, and only the CIF Value and Registration Date
fetch_asycuda_import_export(db_con = NRADWH, cols = c("CIFVALUE", "REGDATE"), dec_type = c("IM4", "IM7"))

# Fetch all results since the beginning of time
fetch_asycuda_import_export(db_con = NRADWH)

}
