% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_asycuda_qa.R
\name{fetch_asycuda_qa}
\alias{fetch_asycuda_qa}
\title{Get ASYCUDA Queries & Amendments Data}
\usage{
fetch_asycuda_qa(db_con, cols = NULL, reg_date = NULL)
}
\arguments{
\item{db_con}{An active DBIConnection object, often created by 'establish_db_connection()'}

\item{cols}{(Optional) A vector or list of columns to include in the subset.}

\item{reg_date}{(Optional) A vector or list where the first element is the start-date and second element the end-date.}
}
\value{
A data.frame containing all the matching observations in the Data Warehouse.
}
\description{
Get ASYCUDA Queries & Amendments Data
}
