% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_asycuda_rmu.R
\name{fetch_asycuda_rmu}
\alias{fetch_asycuda_rmu}
\title{Retrieve ASYCUDA Risk Management Unit Records from the Data Warehouse}
\usage{
fetch_asycuda_rmu(db_con, cols = NULL, reg_date = NULL, dec_type = NULL)
}
\arguments{
\item{db_con}{An active \code{DBIConnection} object created by \code{MRPtools::establish_db_connection()} or \code{DBI::dbConnect}}

\item{cols}{(Optional) A vector or list of columns to include in the subset.}

\item{reg_date}{(Optional) A vector or list where the first element is the start-date and second element the end-date.}

\item{dec_type}{(Optional) A vector or list of Tax Types to filter on. Options are:}
}
\value{
A data.frame containing all the matching observations in the Data Warehouse.
}
\description{
This helper function constructs and executes a query of the \code{ASY_RMU_IMPORT_EXPORT_STATS} table in the Data Warehouse based on the supplied arguments.
}
\examples{
# Establish database connection
# NRADWH <- establish_db_connection(database = "NRADWH")

# Fetch RMU results for 2022
# fetch_asycuda_rmu(db_con = NRADWH, reg_date = c("2022-01-01", "2022-12-31"))

# Fetch results for IM4 and IM7, and only the CIF Value and Registration Date
# fetch_asycuda_rmu(db_con = NRADWH, cols = c("CIF", "REGDATE"), dec_type = c("IM4", "IM7"))

# Fetch all results since the beginning of time
# fetch_asycuda_rmu(db_con = NRADWH)

}
