% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_itas_payments.R
\name{fetch_itas_payments}
\alias{fetch_itas_payments}
\title{Retrieve ITAS Payments from the Data Warehouse}
\usage{
fetch_itas_payments(db_con, payment_date = NULL, office_code = NULL)
}
\arguments{
\item{db_con}{An active \code{DBIConnection} object created by \code{establish_db_connection()} or \code{DBI::dbConnect}}

\item{payment_date}{(Optional) A vector or list where the first element is the start-date and second element the end-date}

\item{office_code}{(Optional) A vector or list of Office Codes to filter on. Options are given by \code{MRPtools::TAX_OFFICE_CODES}}
}
\value{
A data.frame containing all the matching observations in the Data Warehouse
}
\description{
This helper function constructs and executes a query of the \code{DAILY_PAYMENT_REPORT} table joined with \code{REGISTERED_TAXPAYERS_VW} to obtain the office code table.
}
\examples{

# Connect to NRA's data warehouse
NRADWH <- establish_db_connection("NRADWH")

# Retrieve All Payments to Extractive Industries Revenue Unit in 2022
eiru_2022 <- get_itas_payments(NRADWH, payment_date = c("2022-01-01", "2022-12-31"), office_code = "EIRU")

}
