% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_itas_returns.R
\name{fetch_itas_returns}
\alias{fetch_itas_returns}
\title{fetch_itas_returns}
\usage{
fetch_itas_returns(
  db_con,
  document_type = NULL,
  tax_period = NULL,
  due_date = NULL
)
}
\arguments{
\item{db_con}{An active DBIConnection object, often created by 'establish_db_connection()'}

\item{document_type}{A vector or list of Document Types to filter on. Options are: c("PITReturnFinal", "PITReturnProvisional", "GSTReturn", "WHTReturn", "FTTReturn", "CITReturnProvisional", "CITReturnFinal", "PAYEReturn", "PTReturn", "ETReturn")}

\item{tax_period}{(Optional) A vector or list where the first element is the start-date and second element the end-date}

\item{due_date}{(Optional) A vector or list where the first element is the start-date and second element the end-date}
}
\value{
A data.frame containing all the matching observations in the Data Warehouse
}
\description{
fetch_itas_returns
}
