% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_itas_payments.R
\name{get_itas_payments}
\alias{get_itas_payments}
\title{Title}
\usage{
get_itas_payments(db_con, payment_date = NULL, office_code = NULL)
}
\arguments{
\item{db_con}{An active DBIConnection object, often created by 'establish_db_connection()'}

\item{payment_date}{(Optional) A vector or list where the first element is the start-date and second element the end-date}

\item{office_code}{(Optional) A vector or list of Office Codes to filter on. Options are: c("BOSTO", "FTWSTO3", "FTESTO", "KESTO", "SLMTO", "SLLTO", "MSTO", "CIB", "EIRU", "KNSTO", "FTCSTO", "WITO")}
}
\value{
A data.frame containing all the matching observations in the Data Warehouse
}
\description{
Title
}
