% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_directory.R
\name{load_directory}
\alias{load_directory}
\title{load_directory}
\usage{
load_directory(dir, type = "excel", sheets = "all", delim = NULL, ...)
}
\arguments{
\item{dir}{The directory to load.}

\item{type}{Type of files within the supplied directory. Either "excel" or "csv".}

\item{sheets}{(Optional) A numeric vector of indices or a character vector of names for the sheets to use.}

\item{delim}{(Optional) Required if type is 'delim'. Single character used to separate fields within a record.}

\item{...}{(Optional) To pass additional arguments to read_excel().}
}
\value{
A tibble containing the merged observations from the selected files and sheets.
}
\description{
Read and append all files of type 'excel', 'csv', or 'delim'. It uses readr::read_csv() or readxl::read_excel() under the hood. Optionally, if 'type' is "Excel", you can specify which sheets should be loaded. Use ... to pass additional arguments to 'readr' or 'readxl'. Type 'csv' is a convenience wrapper and calls the same readr::read_delim() as does type 'delim'. Using type 'delim' requires that you additionally specify a value for delim. When specifying the type as 'delim', any file format that can be parsed by readr::read_delim() is accepted.
}
