% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_excel_directory.R
\name{load_excel_directory}
\alias{load_excel_directory}
\title{load_excel_directory}
\usage{
load_excel_directory(dir, sheets = "all", quiet = FALSE, ...)
}
\arguments{
\item{dir}{The directory in which to load all files.}

\item{sheets}{A numeric vector of indices or a character vector of names for the sheets to use.}

\item{quiet}{Silences printing progress to the console.}

\item{...}{(Optional) To pass additional arguments to read_excel().}
}
\value{
A tibble containing the merged observations from the selected sheets.
}
\description{
Read and append all Excel files in a directory. Optionally, you can specify which sheets should be loaded using a vector of indices of sheet names. Use ... to pass additional arguments to readxl::read_excel(), which is used under the hood.
}
