% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_database_access.R
\name{setup_database_access}
\alias{setup_database_access}
\title{setup_database_access}
\usage{
setup_database_access(database, host, driver, port = NULL)
}
\arguments{
\item{database}{The name of the database to connect to.}

\item{host}{The I.P. address of the server hosting the DB.}

\item{driver}{The database provider. Either 'mysql' or 'postgres'.}

\item{port}{(Optional) The port number on the server hosting the DB.}
}
\value{
Stores the parameters required for configuring a database connection using keyring::key_set()
}
\description{
setup_database_access
}
