% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/establish_db_connection.R
\name{establish_db_connection}
\alias{establish_db_connection}
\title{Establish a Database Connection}
\usage{
establish_db_connection(database = "NRADWH")
}
\arguments{
\item{database}{The name of the database connection to establish. It must match the name given to 'setup_database_access()'}
}
\value{
An object returned by DBI::dbConnect() which can be used for making subsequent queries against the defined database.
}
\description{
Establish a Database Connection
}
\examples{
# Store credentials for the NRA Data Warehouse (note the IP address may change -- ask IT)
# This only needs to be done once.
setup_database_access(database = "NRADWH", host = "10.18.0.22", driver = RPostgres::Postgres())
NRADWH <- establish_db_connection("NRADWH")

# Subsequently, all you need to do to establish a connection is:
NRADWH <- establish_db_connection("NRADWH")

# Note: the database name in 'setup_database_acess' and 'establish_db_connection' must be the same

}
