% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_asycuda_qa.R
\name{fetch_asycuda_qa}
\alias{fetch_asycuda_qa}
\title{Retrieve ASYCUDA Queries & Amendments Records from Data Warehouse}
\usage{
fetch_asycuda_qa(db_con, cols = NULL, reg_date = NULL)
}
\arguments{
\item{db_con}{An active \code{DBIConnection} object created by \code{MRPtools::establish_db_connection()} or \code{DBI::dbConnect}}

\item{cols}{(Optional) A vector or list of columns to include in the subset.}

\item{reg_date}{(Optional) A vector or list where the first element is the start-date and second element the end-date.}
}
\value{
A data.frame containing all the matching observations in the Data Warehouse.
}
\description{
This helper function constructs and executes a query of the \code{ASY_QAA_MODIFIED_DECLARATIONS} table in the Data Warehouse based on the supplied arguments.
}
\examples{
# Establish database connection
# NRADWH <- establish_db_connection(database = "NRADWH")

# Fetch Import-Export results for 2022
# fetch_asycuda_qa(db_con = NRADWH, reg_date = c("2022-01-01", "2022-12-31"))

# Fetch results for IM4 and IM7, and only the Office Code and Fraud Name 1
# fetch_asycuda_qa(db_con = NRADWH, cols = c("OFFICECODE", "FRAUDNAME1"), dec_type = c("IM4", "IM7"))

# Fetch all results since the beginning of time
# fetch_asycuda_qa(db_con = NRADWH)

}
