% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_itas_account_statement.R
\name{fetch_itas_account_statement}
\alias{fetch_itas_account_statement}
\title{fetch_itas_account_statement}
\usage{
fetch_itas_account_statement(db_con, tax_type = NULL, transaction_date = NULL)
}
\arguments{
\item{db_con}{An active DBIConnection object, often created by 'establish_db_connection()'}

\item{tax_type}{(Optional) A vector or list of Tax Types to filter on. Options are: c("Excise Tax", "Foreign Travel Tax", "Rental Income Tax", "Personal Income Tax", "Suspense Account", "Goods and Services", "Pay As You Earn", "Withholding Tax", "Payroll Tax", "Capital Gains Tax", "Company Income Tax")}

\item{transaction_date}{(Optional) A vector or list where the first element is the start-date and second element the end-date}
}
\value{
A data.frame containing all the matching observations in the Data Warehouse.
}
\description{
fetch_itas_account_statement
}
