% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_itas_returns.R
\name{fetch_itas_returns}
\alias{fetch_itas_returns}
\title{Retrieve ITAS Returns from Data Warehouse}
\usage{
fetch_itas_returns(
  db_con,
  document_type = NULL,
  tax_period = NULL,
  due_date = NULL
)
}
\arguments{
\item{db_con}{An active \code{DBIConnection} object created by \code{establish_db_connection()} or \code{DBI::dbConnect}}

\item{document_type}{A vector or list of Document Types to filter on. Options are given by \code{MRPtools::ITAS_RETURN_TYPES}.}

\item{tax_period}{(Optional) A vector or list where the first element is the start-date and second element the end-date}

\item{due_date}{(Optional) A vector or list where the first element is the start-date and second element the end-date}
}
\value{
A data.frame containing all the matching observations in the Data Warehouse
}
\description{
This helper function constructs and executes a query of the \code{NON_COMPLIANT_TAXPAYERS_VW} table in the Data Warehouse based on the supplied arguments.
}
\examples{
# Establish a database connection
NRADWH <- establish_db_connection("NRADWH")

# Retrieve all PIT Returns
pit_returns <- fetch_itas_returns(NRADWH, document_type = c("PITReturnProvisional", "PITReturnFinal"))

# Retrieve GST returns for Q3 2023
gst_returns_q3 <- fetch_itas_returns(NRADWH, document_type = "Goods and Services", tax_period = c("2023-07-01", "2023-09-30"))

}
