% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_excel.R
\name{load_excel}
\alias{load_excel}
\title{load_excel}
\usage{
load_excel(path, sheets = "all", ...)
}
\arguments{
\item{path}{The Excel file to load.}

\item{sheets}{(Optional) A numeric vector of indices or a character vector of names for the sheets to use.}

\item{...}{(Optional) To pass additional arguments to read_excel().}
}
\value{
A tibble containing the merged observations from the selected sheets.
}
\description{
Read and append sheets within an Excel file. Optionally, you can specify which sheets should be loaded using a vector of indices of sheet names. Use ... to pass additional arguments to readxl::read_excel(), which is used under the hood.
}
