% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_database_access.R
\name{setup_database_access}
\alias{setup_database_access}
\title{Configure Access to NRA Databases}
\usage{
setup_database_access(database, host, driver, port = NULL)
}
\arguments{
\item{database}{The name of the database to connect to.}

\item{host}{The I.P. address of the server hosting the DB.}

\item{driver}{The database provider. Either 'mysql' or 'postgres'.}

\item{port}{(Optional) The port number on the server hosting the DB.}
}
\value{
Stores the parameters required for configuring a database connection using keyring::key_set()
}
\description{
This function allows you to store configurations and credentials for databases using your OS's native secure key storage. When run, it will prompt you to enter your username and password. It also stores details essential for connecting to the database, like the database name, database provider (typically PostgreSQL or MySQL), IP address, and a port number.
}
\examples{
# Store credentials for the NRA Data Warehouse (note the IP address may change -- ask IT)
# This only needs to be done once.
setup_database_access(database = "NRADWH", host = "10.18.0.22", driver = RPostgres::Postgres())
NRADWH <- establish_db_connection("NRADWH")

# Subsequently, all you need to do to establish a connection is:
NRADWH <- establish_db_connection("NRADWH")

# Note: the database name in 'setup_database_acess' and 'establish_db_connection' must be the same

}
